Fix casts

+++ module6/assignment.cu
@@ -51,9 +51,9 @@ __device__ void regTest(uint8_t* a, uint8_t *b, uint32_t *c)
 
     uint8_t ra = (uint8_t) a[tid];
     uint8_t rb = (uint8_t) b[tid];
-    uint32_t plus = (uint32_t) (ra + rb) << c_MSB;
-    uint32_t sub  = (uint32_t) (ra - rb) << c_MLSB;
-    uint32_t mult = (uint32_t) (ra * rb) << c_LMSB;
+    uint32_t plus = ((uint32_t) (ra + rb)) << c_MSB;
+    uint32_t sub  = ((uint32_t) (ra - rb)) << c_MLSB;
+    uint32_t mult = ((uint32_t) (ra * rb)) << c_LMSB;
     uint32_t mod  = (uint32_t) (ra % rb);
     uint32_t res = plus || sub || mult || mod; 
     c[tid] = res;
@@ -65,10 +65,10 @@ __device__ void antiRegTest(uint8_t* a, uint8_t *b, uint32_t *c)
 {
     const int tid = (blockIdx.x * blockDim.x) + threadIdx.x;
     c[tid] = 
-       (uint32_t) ((a[tid] + b[tid]) << c_MSB  ||
-                  (a[tid] - b[tid]) << c_MLSB ||
-                  (a[tid] * b[tid]) << c_LMSB ||
-                  (a[tid] % b[tid]));
+       ((uint32_t)(a[tid] + b[tid]) << c_MSB  ||
+       ((uint32_t)(a[tid] - b[tid]) << c_MLSB) ||
+       ((uint32_t)(a[tid] * b[tid]) << c_LMSB) ||
+       ((uint32_t)(a[tid] % b[tid])));
 }
 
 __global__ void MultKernel(uint8_t *a, uint8_t *b, uint32_t *c, int mode){
